Final Create Api

+++ Backend/ms-accounts/src/main/java/com/bankw/ms_accounts/services/AccountService.java
@@ -1,8 +1,10 @@
 package com.bankw.ms_accounts.services;
 
+import java.math.BigDecimal;
+
 public interface AccountService {
 
     boolean createAccount(String email);
 
-    boolean updateBalance(String username, Double amount);
+    boolean updateBalance(String username, BigDecimal amount);
 }
